requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
ics==0.7.2
python-dateutil==2.8.2
pytest==7.4.3
